Some fixes in GameModel and GameViewModel. Tests for GameModel.

+++ TowerDefence/TowerDefenceGame_LPB/ViewModel/GameViewModel.cs
@@ -119,7 +119,7 @@ namespace TowerDefenceGame_LPB.ViewModel
         private void SetupText()
         {
             Money = model.CurrentPlayer.Money;
-            Round = model.Round;
+            Round = (int)model.Round;
             if (model.Phase % 3 == 0)
             {
                 AdvanceEnable = false;
